optimize stacktraces
The addition of improved stacktraces caused a ~3x slowdown. Profiling
revealed that this was caused by computing the name of iridium functions
for stacktraces. Except in cases where the stacktrace is displayed, this
information is useless and will be thrown away.

Optimize stacktraces by only generating function names when a stacktrace
is shown. This returns performance to that of before adding stacktraces.